// plugins/65xx/src/instructions/variant-tables.generated.ts
export type Ca65Handler =
  | "PutAll"
  | "PutBitBranch"
  | "PutBitBranch_m740"
  | "PutBlockTransfer"
  | "PutLDM_m740"
  | "PutPCRel8"
  | "PutPCRel4510"
  | "Put4510"
  | "Put45GS02"
  | "Put45GS02_Q"
  | "PutJMP"
  | "PutJSR_m740"
  | "PutTAMn"
  | "PutTMA"
  | "PutTMAn"
  | "PutTST";
export type Ca65InstructionRow = readonly [
  mnemonic: string,
  modes: number,
  base: number,
  eaTable: number,
  handler: Ca65Handler,
];

export const ca65EaTable: readonly (readonly number[])[] = [
  [
    0, 0, 5, 13, 15, 21, 29, 31, 0, 25, 18, 0, 7, 17, 23, 1, 0, 0, 0, 3, 19, 9, 0, 9, 0, 0, 0, 0, 0,
    0, 18, 0,
  ],
  [
    8, 8, 4, 12, 0, 20, 28, 0, 20, 28, 0, 128, 0, 16, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 128, 0, 0,
    0, 0, 0,
  ],
  [
    0, 0, 36, 44, 15, 52, 60, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 137, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0,
  ],
  [
    58, 58, 198, 206, 0, 214, 222, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0,
  ],
  [
    26, 26, 230, 238, 0, 246, 254, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0,
  ],
  [
    0, 0, 96, 152, 0, 112, 158, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0,
  ],
  [
    0, 0, 0, 0, 16, 0, 0, 0, 0, 0, 0, 32, 0, 0, 0, 0, 48, 0, 0, 0, 0, 0, 0, 0, 0, 0, 144, 0, 0, 0,
    0, 0,
  ],
  [
    0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 220, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0,
  ],
  [
    0, 64, 1, 65, 0, 9, 73, 0, 0, 0, 0, 81, 0, 0, 0, 0, 0, 0, 0, 0, 0, 64, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0,
  ],
  [
    0, 0, 0, 16, 0, 32, 48, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0,
  ],
  [
    234, 0, 4, 12, 0, 20, 28, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 128, 0, 0, 0, 0, 0, 0,
    0, 0,
  ],
  [
    8, 8, 4, 12, 0, 20, 28, 0, 20, 28, 0, 128, 0, 16, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 0, 0, 128, 0, 0,
    0, 0, 0,
  ],
  [
    0, 0, 0, 76, 0, 0, 0, 0, 0, 0, 178, 108, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0,
  ],
  [
    0, 0, 5, 13, 0, 0, 0, 0, 0, 0, 18, 0, 18, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 18,
    0,
  ],
  [
    10, 10, 6, 14, 0, 22, 30, 0, 0, 0, 18, 0, 18, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 10,
  ],
  [
    3, 3, 4, 12, 0, 20, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    3,
  ],
];